requirements.txt
pyyaml
requests
watchdog
